+++ src/main/java/com/turngo/turngo/repositories/HorarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HorarioRepository extends JpaRepository<Horario, Long> {
+public interface HorarioRepository extends JpaRepository<Horario, Integer> {
 }
